+++ components/marionette/spec/javascripts/itemView.spec.js
@@ -225,15 +225,15 @@ describe("item view", function(){
       view.render();
 
       spyOn(view, "remove").andCallThrough();
-      spyOn(view, "unbindAll").andCallThrough();
+      spyOn(view, "stopListening").andCallThrough();
       spyOn(view, "modelChange").andCallThrough();
       spyOn(view, "collectionChange").andCallThrough();
       spyOn(view, "onBeforeClose").andCallThrough();
       spyOn(view, "onClose").andCallThrough();
       spyOn(view, "trigger").andCallThrough();
 
-      view.bindTo(model, "change:foo", view.modelChange);
-      view.bindTo(collection, "foo", view.collectionChange);
+      view.listenTo(model, "change:foo", view.modelChange);
+      view.listenTo(collection, "foo", view.collectionChange);
 
       view.close();
 
@@ -250,7 +250,7 @@ describe("item view", function(){
     });
 
     it("should unbind any listener to custom view events", function(){
-      expect(view.unbindAll).toHaveBeenCalled();
+      expect(view.stopListening).toHaveBeenCalled();
     });
 
     it("should remove the view's EL from the DOM", function(){
@@ -281,7 +281,7 @@ describe("item view", function(){
         template: "#item-with-checkbox",
 
         setupHandler: function(){
-          this.bindTo(this.model, "change:done", this.render, this);
+          this.listenTo(this.model, "change:done", this.render, this);
         },
 
         events: {
@@ -326,61 +326,6 @@ describe("item view", function(){
       it("should render the view 3 times total", function(){
         expect(spy.callCount).toBe(3);
       });
-    });
-
-  });
-
-  describe("when an item view has a ui elements hash", function() {
-    describe("accessing a ui element from the hash", function() {
-
-      var View = Backbone.Marionette.ItemView.extend({
-        template: "#item-with-checkbox",
-
-        ui: {
-          checkbox: "#chk",
-          unfoundElement: "#not_found"
-        }
-      });
-
-      var view, chk, model;
-
-      beforeEach(function() {
-        loadFixtures("itemWithCheckbox.html");
-
-        model = new Backbone.Model({
-          done: false
-        });
-
-        view = new View({
-          model: model
-        });
-
-        view.render();
-      });
-
-
-      it("should return its jQuery selector if it can be found", function(){
-        expect(view.ui.checkbox.attr("type")).toEqual("checkbox");
-      });
-
-      it("should return an empty jQuery object if it cannot be found", function() {
-        expect(view.ui.unfoundElement.length).toEqual(0);
-      });
-
-      it("should return an up-to-date selector on subsequent renders", function() {
-        // asserting state before subsequent render
-        expect(view.ui.checkbox.attr("checked")).toBeUndefined();
-
-        // setting the model "done" attribute to true will cause the "checked" attribute
-        // to be added to the checkbox element in the subsequent render.
-        view.model.set('done', true);
-        view.render();
-
-        // since the ui elements selectors are refreshed after each render then the associated selector
-        // should point to the newly rendered checkbox element that has the "checked" attribute.
-        expect(view.ui.checkbox.attr("checked")).toBeDefined();
-      });
-
     });
 
   });