+++ components/marionette/spec/javascripts/view.spec.js
@@ -9,7 +9,7 @@ describe("base view", function(){
 
       var view = Backbone.Marionette.View.extend({
         initialize: function(){
-          this.bindTo(this.model, "foo", fooHandler);
+          this.listenTo(this.model, "foo", fooHandler);
         }
       });
 
@@ -27,97 +27,14 @@ describe("base view", function(){
     });
   });
 
-  describe("when a view has string-based model and collection event configuration", function(){
-    var view;
-
-    var View = Backbone.Marionette.View.extend({
-      modelEvents: { 'model-event': 'modelEventHandler modelEventHandler2' },
-      collectionEvents: { 'collection-event': 'collectionEventHandler collectionEventHandler2' },
-
-      modelEventHandler: jasmine.createSpy("model event handler"),
-      collectionEventHandler: jasmine.createSpy("collection event handler"),
-      modelEventHandler2: jasmine.createSpy("model event handler2"),
-      collectionEventHandler2: jasmine.createSpy("collection event handler2")
-    });
-
-    beforeEach(function(){
-      view = new View({
-        model: new Backbone.Model(),
-        collection: new Backbone.Collection()
-      });
-    });
-
-    it("should wire up model events", function(){
-      view.model.trigger("model-event");
-      expect(view.modelEventHandler).toHaveBeenCalled();
-      expect(view.modelEventHandler2).toHaveBeenCalled();
-    });
-
-    it("should wire up collection events", function(){
-      view.collection.trigger("collection-event");
-      expect(view.collectionEventHandler).toHaveBeenCalled();
-      expect(view.collectionEventHandler2).toHaveBeenCalled();
-    });
-
-  });
-
-  describe("when a view has function-based model and collection event configuration", function(){
-    var view;
-
-    var View = Backbone.Marionette.View.extend({
-      modelEvents: { 
-        'model-event': jasmine.createSpy("model event handler")
-      },
-      collectionEvents: { 
-        'collection-event': jasmine.createSpy("collection event handler")
-      }
-    });
-
-    beforeEach(function(){
-      view = new View({
-        model: new Backbone.Model(),
-        collection: new Backbone.Collection()
-      });
-    });
-
-    it("should wire up model events", function(){
-      view.model.trigger("model-event");
-      expect(view.modelEvents['model-event']).toHaveBeenCalled();
-    });
-
-    it("should wire up collection events", function(){
-      view.collection.trigger("collection-event");
-      expect(view.collectionEvents['collection-event']).toHaveBeenCalled();
-    });
-
-  });
-
-  describe("when a view has model event config with a specified handler method that doesn't exist", function(){
-    var getBadViewInstance;
-
-    var View = Backbone.Marionette.View.extend({
-      modelEvents: { "foo": "does_not_exist" }
-    });
-
-    beforeEach(function(){
-      getBadViewInstance = function(){
-        new View({ model: {} })
-      }
-    });
-
-    it("should error when method doesn't exist", function(){
-      expect(getBadViewInstance).toThrow("Method 'does_not_exist' was configured as an event handler, but does not exist.");
-    });
-  });
-
-  describe("when using bindTo for the 'close' event on itself, and closing the view", function(){
+  describe("when using listenTo for the 'close' event on itself, and closing the view", function(){
     var close;
 
     beforeEach(function(){
       close = jasmine.createSpy("close");
-      var view = new Marionette.View();
 
-      view.bindTo(view, "close", close);
+      var view = new Marionette.View();
+      view.listenTo(view, "close", close);
 
       view.close();
     });
@@ -239,4 +156,5 @@ describe("base view", function(){
       expect(view.isClosed).toBe(true);
     });
   });
+
 });