+++ components/marionette/src/marionette.bindEntityEvents.js
@@ -1,10 +1,10 @@
-// Marionette.bindEntityEvents
+// Marionette.bindEntityEvents & unbindEntityEvents
 // ---------------------------
 //
-// This method is used to bind a backbone "entity" (collection/model) 
+// These methods are used to bind/unbind a backbone "entity" (collection/model) 
 // to methods on a target object. 
 //
-// The first paremter, `target`, must have a `bindTo` method from the
+// The first paremter, `target`, must have a `listenTo` method from the
 // EventBinder object.
 //
 // The second parameter is the entity (Backbone.Model or Backbone.Collection)
@@ -13,7 +13,9 @@
 // The third parameter is a hash of { "event:name": "eventHandler" }
 // configuration. Multiple handlers can be separated by a space. A
 // function can be supplied instead of a string handler name. 
-Marionette.bindEntityEvents = (function(){
+
+(function(Marionette){
+  "use strict";
 
   // Bind the event to handlers specified as a string of
   // handler names on the target object
@@ -27,29 +29,62 @@ Marionette.bindEntityEvents = (function(){
         throw new Error("Method '"+ methodName +"' was configured as an event handler, but does not exist.");
       }
 
-      target.bindTo(entity, evt, method, target);
+      target.listenTo(entity, evt, method, target);
     });
   }
 
   // Bind the event to a supplied callback function
   function bindToFunction(target, entity, evt, method){
-      target.bindTo(entity, evt, method, target);
+      target.listenTo(entity, evt, method, target);
   }
 
-  // Export the bindEntityEvents method
-  return function(target, entity, bindings){
+  // Bind the event to handlers specified as a string of
+  // handler names on the target object
+  function unbindFromStrings(target, entity, evt, methods){
+    var methodNames = methods.split(/\s+/);
+
+    _.each(methodNames,function(methodName) {
+      var method = target[method];
+      target.stopListening(entity, evt, method, target);
+    });
+  }
+
+  // Bind the event to a supplied callback function
+  function unbindToFunction(target, entity, evt, method){
+      target.stopListening(entity, evt, method, target);
+  }
+
+  
+  // generic looping function
+  function iterateEvents(target, entity, bindings, functionCallback, stringCallback){
     if (!entity || !bindings) { return; }
 
+    // allow the bindings to be a function
+    if (_.isFunction(bindings)){
+      bindings = bindings.call(target);
+    }
+
+    // iterate the bindings and bind them
     _.each(bindings, function(methods, evt){
 
       // allow for a function as the handler, 
       // or a list of event names as a string
       if (_.isFunction(methods)){
-        bindToFunction(target, entity, evt, methods);
+        functionCallback(target, entity, evt, methods);
       } else {
-        bindFromStrings(target, entity, evt, methods);
+        stringCallback(target, entity, evt, methods);
       }
 
     });
+  }
+ 
+  // Export Public API
+  Marionette.bindEntityEvents = function(target, entity, bindings){
+    iterateEvents(target, entity, bindings, bindToFunction, bindFromStrings);
   };
-})();
+
+  Marionette.unbindEntityEvents = function(target, entity, bindings){
+    iterateEvents(target, entity, bindings, unbindToFunction, unbindFromStrings);
+  };
+
+})(Marionette);